+++ applicationinsights.dev.json
@@ -1,6 +1,6 @@
 {
   "role": {
-    "name": "hmpps-template-kotlin"
+    "name": "hmpps-electronic-monitoring-datastore-api"
   },
   "customDimensions": {
     "service.version": "${BUILD_NUMBER}"
