+++ docker-compose.yml
@@ -1,11 +1,11 @@
 version: "3"
 services:
-  hmpps-template-kotlin:
+  hmpps-electronic-monitoring-datastore-api:
     build:
       context: .
     networks:
       - hmpps
-    container_name: hmpps-template-kotlin
+    container_name: hmpps-electronic-monitoring-datastore-api
     ports:
       - "8080:8080"
     healthcheck:
@@ -14,7 +14,7 @@ services:
       - SERVER_PORT=8080
       - HMPPS_AUTH_URL=http://hmpps-auth:8080/auth
       # TODO: Remove this URL and replace with outgoing service URLs
-      - EXAMPLE_URL=http://hmpps-template-kotlin:8080
+      - EXAMPLE_URL=http://hmpps-electronic-monitoring-datastore-api:8080
       - SPRING_PROFILES_ACTIVE=dev
 
   hmpps-auth:
