+++ helm_deploy/hmpps-electronic-monitoring-datastore-api/Chart.yaml
@@ -1,7 +1,7 @@
 apiVersion: v2
 appVersion: '1.0'
 description: A Helm chart for Kubernetes
-name: hmpps-template-kotlin
+name: hmpps-electronic-monitoring-datastore-api
 version: 0.2.0
 dependencies:
   - name: generic-service
