+++ helm_deploy/hmpps-electronic-monitoring-datastore-api/values.yaml
@@ -1,18 +1,18 @@
 generic-service:
-  nameOverride: hmpps-template-kotlin
+  nameOverride: hmpps-electronic-monitoring-datastore-api
   productId: "HMPPS529" # productId for the product that this belongs too, i.e. DPS001, see README.md for details
 
   replicaCount: 4
 
   image:
-    repository: quay.io/hmpps/hmpps-template-kotlin
+    repository: quay.io/hmpps/hmpps-electronic-monitoring-datastore-api
     tag: app_version # override at deployment time
     port: 8080
 
   ingress:
     enabled: true
     host: app-hostname.local # override per environment
-    tlsSecretName: hmpps-template-kotlin-cert
+    tlsSecretName: hmpps-electronic-monitoring-datastore-api-cert
 
   # Environment variables to load into the deployment
   env:
@@ -27,7 +27,7 @@ generic-service:
   #     [name of environment variable as seen by app]: [key of kubernetes secret to load]
 
   namespace_secrets:
-    hmpps-template-kotlin:
+    hmpps-electronic-monitoring-datastore-api:
       APPINSIGHTS_INSTRUMENTATIONKEY: "APPINSIGHTS_INSTRUMENTATIONKEY"
       # Example client registration secrets
       EXAMPLE_API_CLIENT_ID: "TEMPLATE_KOTLIN_API_CLIENT_ID"
@@ -38,4 +38,4 @@ generic-service:
       - internal
 
 generic-prometheus-alerts:
-  targetApplication: hmpps-template-kotlin
+  targetApplication: hmpps-electronic-monitoring-datastore-api
