+++ helm_deploy/values-prod.yaml-z
@@ -1,15 +1,15 @@
 ---
-# Per environment values which override defaults in hmpps-template-kotlin/values.yaml
+# Per environment values which override defaults in hmpps-electronic-monitoring-datastore-api/values.yaml
 
 generic-service:
   ingress:
-    host: template-kotlin.hmpps.service.justice.gov.uk
+    host: electronic-monitoring-datastore-api.hmpps.service.justice.gov.uk
 
   env:
     HMPPS_AUTH_URL: "https://sign-in.hmpps.service.justice.gov.uk/auth"
     # Template kotlin calls out to itself to provide an example of a service call
     # TODO: This should be replaced by a call to a different service, or removed
-    EXAMPLE_API_URL: "https://template-kotlin.hmpps.service.justice.gov.uk"
+    EXAMPLE_API_URL: "https://electronic-monitoring-datastore-api.hmpps.service.justice.gov.uk"
 
 # CloudPlatform AlertManager receiver to route prometheus alerts to slack
 # See https://user-guide.cloud-platform.service.justice.gov.uk/documentation/monitoring-an-app/how-to-create-alarms.html#creating-your-own-custom-alerts
