+++ settings.gradle.kts
@@ -1,1 +1,1 @@
-rootProject.name = "hmpps-template-kotlin"
+rootProject.name = "hmpps-electronic-monitoring-datastore-api"
