+++ src/main/resources/application.yml
@@ -1,10 +1,10 @@
 info.app:
-  name: HMPPS Template Kotlin
+  name: HMPPS Electronic Monitoring Datastore Api
   version: 1.0
 
 spring:
   application:
-    name: hmpps-template-kotlin
+    name: hmpps-electronic-monitoring-datastore-api
   codec:
     max-in-memory-size: 10MB
