+++ src/main/resources/logback-spring.xml
@@ -12,7 +12,7 @@
     </encoder>
   </appender>
 
-  <logger name="uk.gov.justice.digital.hmpps.templatepackagename.HmppsTemplateKotlinKt" additivity="false" level="DEBUG">
+  <logger name="uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.HmppsElectronicMonitoringDatastoreApiKt" additivity="false" level="DEBUG">
     <appender-ref ref="consoleAppender"/>
   </logger>
