+++ src/test/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringdatastoreapi/integration/NotFoundTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.templatepackagename.integration
+package uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.integration
 
 import org.junit.jupiter.api.Test
