+++ src/test/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringdatastoreapi/integration/wiremock/HmppsAuthMockServer.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.templatepackagename.integration.wiremock
+package uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.integration.wiremock
 
 import com.github.tomakehurst.wiremock.WireMockServer
 import com.github.tomakehurst.wiremock.client.WireMock.aResponse
